requirements.txt
requests>=2.28.0
psycopg2-binary>=2.9.3
urllib3>=1.26.0 